+++ packages/HuYaRecorder/src/index.ts
@@ -35,6 +35,7 @@ function createRecorder(opts: RecorderCreateOpts): Recorder {
     qualityRetry: opts.qualityRetry ?? 0,
     state: "idle",
     api: opts.api ?? "auto",
+    formatName: opts.formatName ?? "auto",
 
     getChannelURL() {
       return `https://www.huya.com/${this.channelId}`;
@@ -114,6 +115,7 @@ const checkLiveStatusAndRecord: Recorder["checkLiveStatusAndRecord"] = async fun
       streamPriorities: this.streamPriorities,
       sourcePriorities: this.sourcePriorities,
       api: this.api,
+      formatName: this.formatName,
     });
   } catch (err) {
     this.state = "idle";