+++ packages/liveManager/CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ## 功能
 
-1. 增加虎牙 quality 类型
-2. 增加虎牙 api 参数支持
+1. 增加虎牙 `quality` 类型、`api`，`formatName` 参数支持
 
 # 1.0.1